requirements.txt
discord.py
python-dotenv
asyncpg 
sqlalchemy 
databases[asyncpg]
Unidecode
pandas
python-docx
reportlab
httpx
beautifulsoup4
pyodbc  